refactor: product와 option을 한 패키지 안에서 관리
