+++ src/main/java/gift/product/model/dto/product/CreateProductAdminRequest.java
@@ -1,8 +1,7 @@
-package gift.product.model.dto;
+package gift.product.model.dto.product;
 
 import static gift.util.Utils.NAME_PATTERN;
 
-import gift.product.model.dto.valid.ExcludeKeyword;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -10,11 +9,10 @@ import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
 import org.springframework.lang.Nullable;
 
-public record CreateProductRequest(
+public record CreateProductAdminRequest(
         @NotBlank
         @Size(max = 15, message = "상품 이름은 15자 이하로 입력해주세요.")
         @Pattern(regexp = NAME_PATTERN, message = "사용할 수 없는 특수문자가 포함되어 있습니다.")
-        @ExcludeKeyword(excludeKeyword = "카카오")
         String name,
 
         @NotNull
@@ -22,6 +20,12 @@ public record CreateProductRequest(
         int price,
 
         @Nullable
-        String imageUrl
+        String imageUrl,
+
+        @NotNull
+        Long sellerId,
+
+        @NotNull
+        Long categoryId
 ) {
 }
