+++ src/main/java/gift/product/model/dto/product/Product.java
@@ -1,4 +1,4 @@
-package gift.product.model.dto;
+package gift.product.model.dto.product;
 
 
 import gift.BaseTimeEntity;
