+++ src/main/java/gift/product/service/ProductAdminService.java
@@ -3,8 +3,8 @@ package gift.product.service;
 import gift.category.model.dto.Category;
 import gift.category.service.CategoryService;
 import gift.product.model.ProductRepository;
-import gift.product.model.dto.CreateProductAdminRequest;
-import gift.product.model.dto.Product;
+import gift.product.model.dto.product.CreateProductAdminRequest;
+import gift.product.model.dto.product.Product;
 import gift.user.model.dto.AppUser;
 import gift.user.service.UserService;
 import org.springframework.stereotype.Service;
