+++ src/main/java/gift/product/service/ProductService.java
@@ -7,10 +7,10 @@ import static gift.util.Utils.TUPLE_WISH_COUNT_KEY;
 import gift.category.model.dto.Category;
 import gift.category.service.CategoryService;
 import gift.product.model.ProductRepository;
-import gift.product.model.dto.CreateProductRequest;
-import gift.product.model.dto.Product;
-import gift.product.model.dto.ProductResponse;
-import gift.product.model.dto.UpdateProductRequest;
+import gift.product.model.dto.product.CreateProductRequest;
+import gift.product.model.dto.product.Product;
+import gift.product.model.dto.product.ProductResponse;
+import gift.product.model.dto.product.UpdateProductRequest;
 import gift.user.exception.ForbiddenException;
 import gift.user.model.dto.AppUser;
 import jakarta.persistence.EntityNotFoundException;
