+++ src/test/java/gift/product/ProductAdminServiceTest.java
@@ -11,8 +11,8 @@ import static org.mockito.Mockito.when;
 import gift.category.model.dto.Category;
 import gift.category.service.CategoryService;
 import gift.product.model.ProductRepository;
-import gift.product.model.dto.CreateProductAdminRequest;
-import gift.product.model.dto.Product;
+import gift.product.model.dto.product.CreateProductAdminRequest;
+import gift.product.model.dto.product.Product;
 import gift.product.service.ProductAdminService;
 import gift.product.service.ProductService;
 import gift.user.model.dto.AppUser;
