+++ src/test/java/gift/product/ProductRepositoryTest.java
@@ -9,7 +9,7 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import gift.category.model.CategoryRepository;
 import gift.category.model.dto.Category;
 import gift.product.model.ProductRepository;
-import gift.product.model.dto.Product;
+import gift.product.model.dto.product.Product;
 import gift.user.model.UserRepository;
 import gift.user.model.dto.AppUser;
 import gift.user.model.dto.Role;
