+++ src/test/java/gift/wishlist/WishListRepositoryTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import gift.category.model.CategoryRepository;
 import gift.category.model.dto.Category;
 import gift.product.model.ProductRepository;
-import gift.product.model.dto.Product;
+import gift.product.model.dto.product.Product;
 import gift.user.model.UserRepository;
 import gift.user.model.dto.AppUser;
 import gift.user.model.dto.Role;
